backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]
sqlalchemy==2.0.36
pydantic==2.9.2
pydantic-settings==2.5.2
psycopg2-binary
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv
alembic==1.13.2
redis
rq
bcrypt==4.0.1
email-validator==2.2.0
